Update synth revamp + palette reuse fix (#331)
