+++ x/emissions/keeper/inference_synthesis/synth_palette_bootstrap.go
@@ -1,6 +1,9 @@
 package inference_synthesis
 
-import errorsmod "cosmossdk.io/errors"
+import (
+	errorsmod "cosmossdk.io/errors"
+	emissionstypes "github.com/allora-network/allora-chain/x/emissions/types"
+)
 
 // Bootstraps xRegrets, allxsAreNew (x="inferer"|"forecasters") for the inferers and forecasters in the palette
 // Just requires these props:: ctx, k, topicId, inferers, forecasts
@@ -34,4 +37,49 @@ func (p *SynthPalette) BootstrapRegretData() error {
 	}
 
 	return nil
+}
+
+// Clone creates a deep copy of the SynthPalette.
+func (p SynthPalette) Clone() SynthPalette {
+	inferenceByWorker := make(map[Worker]*emissionstypes.Inference, len(p.inferenceByWorker))
+	for k, v := range p.inferenceByWorker {
+		inferenceByWorker[k] = v
+	}
+	forecastByWorker := make(map[Worker]*emissionstypes.Forecast, len(p.forecastByWorker))
+	for k, v := range p.forecastByWorker {
+		forecastByWorker[k] = v
+	}
+	forecastImpliedInferenceByWorker := make(map[Worker]*emissionstypes.Inference, len(p.forecastImpliedInferenceByWorker))
+	for k, v := range p.forecastImpliedInferenceByWorker {
+		inferenceCopy := *v
+		forecastImpliedInferenceByWorker[k] = &inferenceCopy
+	}
+	infererRegrets := make(map[Worker]StatefulRegret, len(p.infererRegrets))
+	for k, v := range p.infererRegrets {
+		infererRegrets[k] = v
+	}
+	forecasterRegrets := make(map[Worker]StatefulRegret, len(p.forecasterRegrets))
+	for k, v := range p.forecasterRegrets {
+		forecasterRegrets[k] = v
+	}
+
+	return SynthPalette{
+		ctx:                              p.ctx,
+		k:                                p.k,
+		topicId:                          p.topicId,
+		inferers:                         append([]Worker(nil), p.inferers...),
+		inferenceByWorker:                inferenceByWorker,
+		infererRegrets:                   infererRegrets,
+		forecasters:                      append([]Worker(nil), p.forecasters...),
+		forecastByWorker:                 forecastByWorker,
+		forecastImpliedInferenceByWorker: forecastImpliedInferenceByWorker,
+		forecasterRegrets:                forecasterRegrets,
+		allInferersAreNew:                p.allInferersAreNew,
+		allForecastersAreNew:             p.allForecastersAreNew,
+		allWorkersAreNew:                 p.allWorkersAreNew,
+		networkCombinedLoss:              p.networkCombinedLoss,
+		epsilon:                          p.epsilon,
+		pNorm:                            p.pNorm,
+		cNorm:                            p.cNorm,
+	}
 }
