+++ x/emissions/keeper/inference_synthesis/synth_palette_factory.go
@@ -6,15 +6,15 @@ import (
 
 // Could use Builder pattern in the future to make this cleaner
 func (f *SynthPaletteFactory) BuildPaletteFromRequest(req SynthRequest) SynthPalette {
-	inferenceByWorker := MakeMapFromInfererToTheirInference(req.inferences.Inferences)
-	forecastByWorker := MakeMapFromForecasterToTheirForecast(req.forecasts.Forecasts)
+	inferenceByWorker := MakeMapFromInfererToTheirInference(req.Inferences.Inferences)
+	forecastByWorker := MakeMapFromForecasterToTheirForecast(req.Forecasts.Forecasts)
 	sortedInferers := alloraMath.GetSortedKeys(inferenceByWorker)
 
 	// Those values not from req are to be considered defaults
 	palette := SynthPalette{
-		ctx:                              req.ctx,
-		k:                                req.k,
-		topicId:                          req.topicId,
+		ctx:                              req.Ctx,
+		k:                                req.K,
+		topicId:                          req.TopicId,
 		inferers:                         sortedInferers,
 		inferenceByWorker:                inferenceByWorker,
 		infererRegrets:                   make(map[string]StatefulRegret), // Populated below
@@ -25,19 +25,19 @@ func (f *SynthPaletteFactory) BuildPaletteFromRequest(req SynthRequest) SynthPal
 		allInferersAreNew:                true,                            // Populated below
 		allForecastersAreNew:             true,                            // Populated below
 		allWorkersAreNew:                 true,                            // Populated below
-		networkCombinedLoss:              req.networkCombinedLoss,
-		epsilon:                          req.epsilon,
-		pNorm:                            req.pNorm,
-		cNorm:                            req.cNorm,
+		networkCombinedLoss:              req.NetworkCombinedLoss,
+		epsilon:                          req.Epsilon,
+		pNorm:                            req.PNorm,
+		cNorm:                            req.CNorm,
 	}
-
-	// Populates: infererRegrets, forecasterRegrets, allInferersAreNew, allForecastersAreNew
-	palette.BootstrapRegretData()
-	palette.allWorkersAreNew = palette.allInferersAreNew && palette.allForecastersAreNew
 
 	// Populates: forecastImpliedInferenceByWorker, forecasters
 	palette.UpdateForecastImpliedInferences()
 	palette.forecasters = alloraMath.GetSortedKeys(palette.forecastImpliedInferenceByWorker)
+
+	// Populates: infererRegrets, forecasterRegrets, allInferersAreNew, allForecastersAreNew
+	palette.BootstrapRegretData()
+	palette.allWorkersAreNew = palette.allInferersAreNew && palette.allForecastersAreNew
 
 	return palette
 }
